Move MemoryManager to new package
